chore: refactor to AppZeroState, create PermissionsProvider
